# console/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brand Console API",
    "version": "0.1.0",
    "description": "The local Express API behind the brand console. Runs unauthenticated on http://localhost:5174. Read-only reference; the console never holds a secret because there is no secret to hold."
  },
  "servers": [{ "url": "http://localhost:5174", "description": "Local dev server" }],
...
  "paths": {
...
    "/api/queue": {
      "get": {
        "tags": ["Queue"],
        "summary": "List idea-queue items, scored desc",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "enum": ["new", "seeded", "drafting", "drafted", "archived"] },
            "description": "Optional status filter."
          }
        ],
        "responses": {
          "200": {
            "description": "The idea queue.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/IdeaQueueItem" } }
              }
            }
          }
        }
      }
    },
    "/api/queue/{id}/seed": {
      "post": {
        "tags": ["Queue"],
        "summary": "Attach a seed and mark the item seeded",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "seed": { "type": "string" } },
                "required": ["seed"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated queue item.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdeaQueueItem" } } }
          }
        }
      }
    },
    "/api/drafts": {
      "get": {
        "tags": ["Drafts"],
        "summary": "List drafts (newest first)",
        "responses": {
          "200": {
            "description": "The draft list.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Draft" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Drafts"],
        "summary": "Create a draft from an idea",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ideaId": { "type": "string" },
                  "hookOptions": { "type": "array", "items": { "type": "string" } },
                  "body": { "type": "string" },
                  "close": { "type": "string" },
                  "mediaSuggestion": { "type": "string" }
                },
                "required": ["ideaId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created draft.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Draft" } } }
          }
        }
      }
    },
    "/api/drafts/{id}": {
      "get": {
        "tags": ["Drafts"],
        "summary": "One draft",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "The draft.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Draft" } } }
          },
          "404": { "description": "No such draft." }
        }
      },
      "patch": {
        "tags": ["Drafts"],
        "summary": "Partial edit of a draft",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hookOptions": { "type": "array", "items": { "type": "string" } },
                  "body": { "type": "string" },
                  "close": { "type": "string" },
                  "mediaSuggestion": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated draft.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Draft" } } }
          }
        }
      }
    },
    "/api/drafts/{id}/review": {
      "post": {
        "tags": ["Drafts"],
        "summary": "Set the stored review status",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reviewStatus": { "type": "string", "enum": ["pending", "passed", "failed", "edited"] }
                },
                "required": ["reviewStatus"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated draft.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Draft" } } }
          }
        }
      }
    },
    "/api/drafts/{id}/publish": {
      "post": {
        "tags": ["Drafts"],
        "summary": "Publish a draft into the post archive",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "permalink": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The published post.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublishedPost" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IdeaQueueItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "pillar": { "type": "string" },
          "tag": { "type": "string", "enum": ["needs-your-take", "ready-to-draft"] },
          "sourceRef": { "type": ["string", "null"] },
          "proposedAngle": { "type": "string" },
          "seed": { "type": ["string", "null"] },
          "score": { "type": "number" },
          "status": { "type": "string", "enum": ["new", "seeded", "drafting", "drafted", "archived"] },
          "createdAt": { "type": "integer", "description": "Epoch milliseconds." }
        }
      },
      "Draft": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "ideaId": { "type": "string" },
          "hookOptions": { "type": "array", "items": { "type": "string" } },
          "body": { "type": "string" },
          "close": { "type": "string" },
          "mediaSuggestion": { "type": "string" },
          "reviewStatus": { "type": "string", "enum": ["pending", "passed", "failed", "edited"] },
          "createdAt": { "type": "integer", "description": "Epoch milliseconds." }
        }
      },
      "PublishedPost": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "draftId": { "type": "string" },
          "permalink": { "type": ["string", "null"] },
          "publishedAt": { "type": "integer", "description": "Epoch milliseconds." },
          "pillar": { "type": ["string", "null"], "description": "Resolved via post -> draft -> idea." }
        }
      },
...
    }
  }
}